many: address review comments
