+++ overlord/snapstate/handlers.go
@@ -2285,40 +2285,6 @@ func daemonRestartReason(st *state.State, typ snap.Type) string {
 	return "Requested daemon restart (snapd snap)."
 }
 
-func (m *SnapManager) doCheckReseal(t *state.Task, _ *tomb.Tomb) error {
-	st := t.State()
-	st.Lock()
-	defer st.Unlock()
-
-	snapsup, err := TaskSnapSetup(t)
-	if err != nil {
-		return err
-	}
-
-	logger.Debugf("finish restart from doCheckReseal")
-	if err := FinishRestart(t, snapsup, FinishRestartOptions{}); err != nil {
-		return err
-	}
-
-	deviceCtx, err := DeviceCtx(st, t, nil)
-	if err != nil {
-		return err
-	}
-	if !deviceCtx.HasModeenv() || !deviceCtx.RunMode() {
-		return nil
-	}
-
-	unlocker := st.Unlocker()
-	st.Unlock()
-	defer st.Lock()
-
-	if err := boot.CheckResealKeyToModeenv(dirs.GlobalRootDir, unlocker); err != nil {
-		return fmt.Errorf("cannot validate key reseal: %v", err)
-	}
-
-	return nil
-}
-
 // maybeDiscardNamespacesOnSnapdDowngrade must be called when we are about to
 // activate a different snapd version. It checks whether we are performing a
 // downgrade to a snapd version that does not support the
